Moved calendar, subscription and base services to separate files, moved utility functions to utils file
