+++ dev/demo-app/webapp/ClientApp/services/graphServices/index.ts
@@ -1,0 +1,7 @@
+import { CalendarService } from './calendarService';
+import { SubscriptionService } from './subscriptionService';
+
+export const instances = {
+    calendar: new CalendarService(),
+    subscription: new SubscriptionService()
+};
